Reverted back to functioning component
